src/main/java/ru/itsjava/service/UserServiceImpl.java
package ru.itsjava.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import ru.itsjava.domain.Pet;
import ru.itsjava.domain.User;
import ru.itsjava.repository.PetRepository;
import ru.itsjava.repository.UserRepository;

import java.util.List;
import java.util.Optional;

@Service
@RequiredArgsConstructor
public class UserServiceImpl implements UserService {
    private final UserRepository userRepository;
    private final PetRepository petRepository;

    @Transactional
    @Override
    public void createUser(User user) {
        Optional<Pet> petByBreed = petRepository.getByBreed(user.getPet().getBreed());
        petByBreed.ifPresent(user::setPet);
        userRepository.save(user);
    }

    @Transactional(readOnly = true)
    @Override
    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    @Transactional
    @Override
    public void updateUser(User user) {
        Optional<Pet> petByBreed = petRepository.getByBreed(user.getPet().getBreed());
        petByBreed.ifPresent(user::setPet);
        userRepository.save(user);
    }

    @Transactional
    @Override
    public void deleteUser(User user) {
        userRepository.delete(user);
    }

    @Transactional(readOnly = true)
    @Override
    public Optional<User> getUserById(long id) {
        return userRepository.findById(id);
    }

}
